Moved GennGeneService into sub folder
