+++ lib/genetically_evolving_neural_network.dart
@@ -18,7 +18,7 @@ part 'package:genetically_evolving_neural_network/services/genn_dna_service.dart
 part 'package:genetically_evolving_neural_network/services/genn_entity_service.dart';
 part 'package:genetically_evolving_neural_network/services/genn_fitness_service.dart';
 part 'package:genetically_evolving_neural_network/services/genn_gene_mutation_service.dart';
-part 'package:genetically_evolving_neural_network/services/genn_gene_service.dart';
+part 'package:genetically_evolving_neural_network/services/genn_gene_service/genn_gene_service.dart';
 part 'package:genetically_evolving_neural_network/services/perceptron_layer_mutation_service.dart';
 
 /// Represents a Genetically Evolving Neural Network.
